Refactor: move image signing logic to api.controller and update dao references
